refactor(server): 拆出 ServerList 分组头、卡片项、表格项与上下文菜单子组件
